cleaned up console.logs and moved the load tests and relevent files to new loadTest directory
