+++ main_process/main_httpController.js
@@ -454,11 +454,6 @@ const httpController = {
       return cookieFormat;
     });
   },
-  // ----------------------------------------------------------------------------
-
-  // runLoadTest(event, reqResObj) {
-  //   event.sender.send('reqResUpdate', reqResObj);
-  // },
 
   // parses SSE format into an object
   // SSE format -> 'key1: value1\nkey2: value2\nkey3: value3\n\n
@@ -476,18 +471,6 @@ const httpController = {
 };
 
 module.exports = () => {
-  // create ipc event listener for load-test
-  ipcMain.on('http-load-test', (event, { id, results }) => {
-    const reqResArr = store.getState().reqRes.reqResArray;
-    const reqResObj = reqResArr.find((el) => el.id === id);
-
-    if (reqResObj) {
-      reqResObj.response.events = results;
-      event.sender.send('reqResUpdate', reqResObj);
-    } else {
-      console.error('Error: reqResObj not found');
-    }
-  });
   // creating our event listeners for IPC events
   ipcMain.on('open-http', (event, reqResObj) => {
     // we pass the event object into these controller functions so that we can invoke event.sender.send when we need to make response to renderer process
